Remove static Crypto property from BaseClient
We have decided that we don’t want the tree-shakable version of the
library to make use of static properties or methods, as they are not
amenable to tree-shaking. There’s also no obvious way to set a static
property in the API that we’ve chosen for tree-shaking (since we pass
dependencies into the class’s constructor).

Instead, the tree-shakable version of the library now exports standalone
functions which replace the removed static methods.